First iteration of anoncreds_types structure
