+++ aries/misc/anoncreds_types/src/data_types/cred_def/private.rs
@@ -1,0 +1,6 @@
+use anoncreds_clsignatures::CredentialPrivateKey;
+
+#[derive(Debug, Deserialize, Serialize)]
+pub struct CredentialDefinitionPrivate {
+    pub value: CredentialPrivateKey,
+}
