+++ aries/misc/anoncreds_types/src/data_types/cred_offer/mod.rs
@@ -1,0 +1,2 @@
+mod public;
+pub use public::*;
