+++ aries/misc/anoncreds_types/src/data_types/cred_request/mod.rs
@@ -1,0 +1,3 @@
+mod public;
+
+pub use public::*;
