+++ aries/misc/anoncreds_types/src/data_types/credential/mod.rs
@@ -1,0 +1,3 @@
+mod private;
+
+pub use private::*;
