+++ aries/misc/anoncreds_types/src/data_types/link_secret/private.rs
