+++ aries/misc/anoncreds_types/src/data_types/rev_reg/mod.rs
@@ -1,0 +1,3 @@
+mod public;
+
+pub use public::*;
