+++ aries/misc/anoncreds_types/src/data_types/rev_reg/public.rs
