+++ aries/misc/anoncreds_types/src/data_types/rev_reg_def/mod.rs
@@ -1,0 +1,5 @@
+mod private;
+mod public;
+
+pub use private::*;
+pub use public::*;
