+++ aries/misc/anoncreds_types/src/data_types/rev_reg_def/private.rs
@@ -1,0 +1,6 @@
+use anoncreds_clsignatures::RevocationKeyPrivate;
+
+#[derive(Debug, Deserialize, Serialize)]
+pub struct RevocationRegistryDefinitionPrivate {
+    pub value: RevocationKeyPrivate,
+}
