+++ aries/misc/anoncreds_types/src/data_types/rev_reg_def/public.rs
@@ -1,9 +1,9 @@
 use std::str::FromStr;
 
-use crate::cl::{RevocationKeyPrivate, RevocationKeyPublic};
+use crate::cl::RevocationKeyPublic;
+use crate::data_types::cred_def::CredentialDefinitionId;
+use crate::data_types::issuer_id::IssuerId;
 use crate::{error::ConversionError, impl_anoncreds_object_identifier};
-
-use super::{cred_def::CredentialDefinitionId, issuer_id::IssuerId};
 
 pub const CL_ACCUM: &str = "CL_ACCUM";
 
@@ -58,9 +58,4 @@ impl Validatable for RevocationRegistryDefinition {
 
         Ok(())
     }
-}
-
-#[derive(Debug, Deserialize, Serialize)]
-pub struct RevocationRegistryDefinitionPrivate {
-    pub value: RevocationKeyPrivate,
 }
