+++ aries/misc/anoncreds_types/src/data_types/schema/mod.rs
@@ -1,0 +1,2 @@
+mod public;
+pub use public::*;
