+++ aries/misc/anoncreds_types/src/data_types/schema/public.rs
@@ -1,8 +1,6 @@
-use crate::impl_anoncreds_object_identifier;
+use crate::{data_types::issuer_id::IssuerId, impl_anoncreds_object_identifier};
 
 use std::collections::HashSet;
-
-use super::issuer_id::IssuerId;
 
 pub const MAX_ATTRIBUTES_COUNT: usize = 125;
