+++ aries/misc/test_utils/src/mockdata/mock_ledger.rs
@@ -11,7 +11,7 @@ use async_trait::async_trait;
 
 use crate::constants::{
     rev_def_json, CRED_DEF_JSON, DEFAULT_AUTHOR_AGREEMENT, REQUEST_WITH_ENDORSER,
-    REV_REG_DELTA_JSON, REV_REG_ID, REV_REG_JSON, SCHEMA_JSON,
+    REV_REG_DELTA_JSON, REV_REG_ID, REV_REG_JSON
 };
 
 #[derive(Debug)]
